requirements.txt
# Backend dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic==1.10.13  # Downgraded for langchain compatibility
python-dotenv>=1.0.0
httpx>=0.25.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
langchain-core>=0.1.30
langchain>=0.1.12
groq>=0.4.2

# Frontend dependencies
streamlit>=1.28.0
uuid>=1.30

# Development dependencies
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.10.1
isort>=5.12.0
flake8>=6.1.0